Strip out 'ConfigurationNotRecognizedError'...
I have no idea (and am unable to recreate) the circumstances that cause an `inspect()` call on a `WorkspaceConfiguration` object to yield `undefined`. Even if I intentionally bungle the section name and child name, I still seem to obtain at least an empty object that is not `undefined`.

So instead of throwing a `ConfigurationNotRecognizedError`, I will just throw a generic error when the problem arises.