Initially there was an attempt to generate automated tests but this was abandoned for a few reasons:
1. The code to generate the tests was pretty much just a replication
of the logic in the actual code we wanted to test.

2. The amount of test cases for VCReaderCompat is 2^14, and since it
takes a while for vscode to set and read configuration values (on the
author's computer it ranged from 50ms to 200ms just for a single
set-read-check run), the total time taken for the tests can be up to
50 minutes in the worst case.

Therefore automated test generation has been removed from the master
branch. This branch merely serves as an archive of the old test
generation code.