Change `prepack` script to `prepare`... [cont.]
The `prepack` script doesn't seem to work any more. The point of the
`prepack` script was that it was supposed to run after the git repository
is downloaded by npm, such that in the case of this package, the code
is compiled before installation.

The new `prepare` script seems to serve that role now.